Extracting helper function for parsing timestamp filters.
